+++ test/jest/console-baseline-unit.json
@@ -496,6 +496,10 @@
     "ui/components/multichain/network-manager/network-manager.test.tsx": {
       "error: Unexpected key \"internalAccounts\" found in": 1
     },
+    "ui/components/multichain/networks-form/networks-form.test.js": {
+      "React: Act warnings (component updates not wrapped)": 1,
+      "React: State updates on unmounted components": 1
+    },
     "ui/components/multichain/notifications-settings-box/notifications-settings-box.test.tsx": {
       "React: componentWill* lifecycle deprecations": 1
     },
@@ -1538,10 +1542,6 @@
     "ui/pages/settings-v2/transactions-tab/transactions-tab.test.tsx": {
       "React: componentWill* lifecycle deprecations": 1,
       "Reselect: Identity function warnings": 1
-    },
-    "ui/pages/settings/networks-tab/networks-form/networks-form.test.js": {
-      "React: Act warnings (component updates not wrapped)": 1,
-      "React: State updates on unmounted components": 1
     },
     "ui/pages/settings/transaction-shield-tab/claims-form/claims-form.test.tsx": {
       "MetaMask: Background connection not initialized": 4,
