+++ ui/components/multichain/asset-picker-amount/asset-picker-modal/AssetList.test.tsx
@@ -56,7 +56,7 @@ const mockUseSafeChains = jest.fn().mockReturnValue({
   safeChains: [],
 });
 jest.mock(
-  '../../../../pages/settings/networks-tab/networks-form/use-safe-chains',
+  '../../../networks-form/use-safe-chains',
   () => ({
     useSafeChains: () => mockUseSafeChains(),
   }),
