+++ ui/components/multichain/multichain-components.scss
@@ -23,6 +23,7 @@
 @import 'network-list-item';
 @import 'network-list-item-menu';
 @import 'network-list-menu';
+@import 'networks-form';
 @import 'carousel';
 @import 'network-list-menu/add-rpc-url-modal';
 @import 'network-list-menu/add-block-explorer-modal';
