+++ ui/components/multichain/network-list-menu/network-list-menu.tsx
@@ -110,8 +110,8 @@ import {
   getCompletedOnboarding,
   getIsUnlocked,
 } from '../../../ducks/metamask/metamask';
-import NetworksForm from '../../../pages/settings/networks-tab/networks-form';
-import { useNetworkFormState } from '../../../pages/settings/networks-tab/networks-form/networks-form-state';
+import NetworksForm from '../networks-form';
+import { useNetworkFormState } from '../networks-form/networks-form-state';
 import { openWindow } from '../../../helpers/utils/window';
 import { endTrace, TraceName } from '../../../../shared/lib/trace';
 import PopularNetworkList from './popular-network-list/popular-network-list';
