+++ ui/components/multichain/network-manager/components/add-network/AddNetwork.tsx
@@ -1,8 +1,8 @@
 import { UpdateNetworkFields } from '@metamask/network-controller';
 import React from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { NetworksForm } from '../../../../../pages/settings/networks-tab/networks-form/networks-form';
-import { useNetworkFormState } from '../../../../../pages/settings/networks-tab/networks-form/networks-form-state';
+import { NetworksForm } from '../../../networks-form/networks-form';
+import { useNetworkFormState } from '../../../networks-form/networks-form-state';
 
 type AddNetworkProps = {
   networkFormState: ReturnType<typeof useNetworkFormState>;
