+++ ui/components/multichain/network-manager/hooks/useNewRPC.ts
@@ -1,7 +1,7 @@
 import { useCallback } from 'react';
 import * as URI from 'uri-js';
 import { RpcEndpointType } from '@metamask/network-controller';
-import { useNetworkFormState } from '../../../../pages/settings/networks-tab/networks-form/networks-form-state';
+import { useNetworkFormState } from '../../networks-form/networks-form-state';
 
 type UseNewRPCProps = {
   rpcUrls: ReturnType<typeof useNetworkFormState>['rpcUrls'];
