+++ ui/components/multichain/networks-form/index.scss
@@ -1,0 +1,15 @@
+.networks-form {
+  &__scrollable {
+    overflow-y: auto;
+  }
+
+  &__footer {
+    position: sticky;
+    bottom: 0;
+    box-shadow: 0 0 8px 0 var(--color-shadow-default);
+
+    &--page {
+      box-shadow: none;
+    }
+  }
+}
