+++ ui/components/multichain/networks-form/use-safe-chains.ts
@@ -2,11 +2,11 @@ import { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { isStrictHexString } from '@metamask/utils';
 
-import { getUseSafeChainsListValidation } from '../../../../selectors';
-import fetchWithCache from '../../../../../shared/lib/fetch-with-cache';
-import { CHAIN_SPEC_URL } from '../../../../../shared/constants/network';
-import { DAY } from '../../../../../shared/constants/time';
-import { hexToDecimal } from '../../../../../shared/lib/conversion.utils';
+import { getUseSafeChainsListValidation } from '../../../selectors';
+import fetchWithCache from '../../../../shared/lib/fetch-with-cache';
+import { CHAIN_SPEC_URL } from '../../../../shared/constants/network';
+import { DAY } from '../../../../shared/constants/time';
+import { hexToDecimal } from '../../../../shared/lib/conversion.utils';
 
 export type SafeChain = {
   chainId: string;
