+++ ui/pages/confirmations/hooks/alerts/useAddEthereumChainAlerts.test.ts
@@ -1,7 +1,7 @@
 import { renderHook } from '@testing-library/react-hooks';
 import { Severity } from '../../../../helpers/constants/design-system';
 import * as confirmContext from '../../context/confirm';
-import * as safeChains from '../../../settings/networks-tab/networks-form/use-safe-chains';
+import * as safeChains from '../../../../components/multichain/networks-form/use-safe-chains';
 import * as rpcUtils from '../../../../../shared/lib/rpc.utils';
 import { useAddEthereumChainAlerts } from './useAddEthereumChainAlerts';
