+++ ui/pages/confirmations/hooks/alerts/useAddEthereumChainAlerts.ts
@@ -6,7 +6,7 @@ import type { Alert } from '../../../../ducks/confirm-alerts/confirm-alerts';
 import { DEPRECATED_NETWORKS } from '../../../../../shared/constants/network';
 import { AddEthereumChainContext } from '../../external/add-ethereum-chain/types';
 import { useI18nContext } from '../../../../hooks/useI18nContext';
-import { useSafeChains } from '../../../settings/networks-tab/networks-form/use-safe-chains';
+import { useSafeChains } from '../../../../components/multichain/networks-form/use-safe-chains';
 import { jsonRpcRequest } from '../../../../../shared/lib/rpc.utils';
 import { RowAlertKey } from '../../../../components/app/confirm/info/row/constants';
