+++ ui/pages/settings/index.scss
@@ -1,7 +1,6 @@
 @use "design-system";
 
 @import '../settings-v2/debug-tab/developer-options-tab/index';
-@import 'networks-tab/index';
 @import 'transaction-shield-tab/index';
 
 .settings-page {
